Apply a32e8a8d74dd0e5da931bd152609662f0c67a35e - Fix an inconsistency in zilch shader tests where names were in hashmap order + previous random changes

+++ Libraries/Common/StringBuilder.cpp
@@ -121,7 +121,7 @@ void StringBuilder::Append(cstr begin, uint sizeInBytes)
 void StringBuilder::Append(Rune rune)
 {
   byte utf8Bytes[4];
-  int bytesRead = UTF8::UnpackUtf8RuneIntoBuffer(rune, utf8Bytes);
+  size_t bytesRead = UTF8::UnpackUtf8RuneIntoBuffer(rune, utf8Bytes);
   ByteBuffer::Append(utf8Bytes, bytesRead);
 }
